moved vendor libs to assets
